portfolio-engine/pyproject.toml
[project]
name = "portfolio-engine"
version = "0.27.2"
requires-python = ">=3.11"
dependencies = [
    "fastapi", "uvicorn[standard]", "sqlalchemy", "aiomysql", "pydantic", "python-dotenv", "cryptography"
]
